[DSCP-125] Fixed diss up to `submitAssignment`, incl.
