+++ .gitignore
@@ -26,3 +26,4 @@ educator-db
 witness-db
 stylish-haskell.yaml
 test.db
+.stylish-haskell.yaml
