+++ hpack/definitions.yaml
@@ -57,10 +57,10 @@ _definitions:
         main:        Main.hs
         source-dirs: tests
 
-        ghc-options: []
-          # - -threaded
-          # - -rtsopts
-          # - -with-rtsopts=-N
+        ghc-options:
+          - -threaded
+          - -rtsopts
+          - -with-rtsopts=-N
 
     - &bench-common
         source-dirs: bench
